Added a "Start" scene for Room1Perfect
This prevents a clone of the girl from loading upon returning to the
room from the dollhouse. Start the game from Room1PerfectStart, then the
rest should work?

I'm in class now lol don't have time atm to figure out how to move or
hide the girl when the dollhouse loads.